[Fix] 오류 & 최종 디자인 수정 (#134)
* [fix]: 사용자 트랜드 수집 오류 수정

* [fix]: 카카오톡 로그인 오류 수정

* [fix]: 뒤로가기 토스트 수정

* [fix]: 검색 오류 수정

* [fix]: 관계 추가 시 이전 데이터 유지

* [feat]: 디자인 QA 적용

* [feat]: 경사 종류 아이콘 교체

* [feat]: 경사 종류 아이콘 교체

* [feat]: 일정 스크린 스티키 적용

* [fix]: 일정 날짜 클릭 시 오늘로 이동

* [fix]: 선택 캘린더 변경

* [fix]

* [fix]

* [fix]

+++ data/src/main/kotlin/ac/dnd/mour/android/data/remote/network/model/statistics/GetGroupStatisticsRes.kt
@@ -20,7 +20,7 @@ data class GetGroupStatisticsItemRes(
     @SerialName("event")
     val event: String,
     @SerialName("amount")
-    val amount: Long
+    val amount: Double
 ) : DataMapper<GroupStatistics> {
     override fun toDomain(): GroupStatistics {
         return GroupStatistics(
